11: Moved form-related files into components folder for better structure
