+++ src/App.jsx
@@ -1,4 +1,4 @@
-import ContactForm from "./ContactForm";
+import ContactForm from "./components/ContactForm";
 
 function App() {
   return (
